prints a buffer

+++ 0x06-pointers_arrays_strings/104-print_buffer.c
@@ -22,7 +22,7 @@ void print_buffer(char *b, int size)
 		for (j = 0; j < 10; j++)
 		{
 			if (i + j < size)
-				printf("%02x", (unsigned char)b[i + j]);
+				printf("%02x", (unsigned int)(unsigned char)b[i + j]);
 			else
 				printf("  ");
 			if (j % 2 == 1 && j != 9)
